refactor: regionSearch 컴포넌트 구현, 로직 커스텀 훅으로 분리
