+++ frontend/src/types/region.ts
@@ -2,3 +2,8 @@ export interface RegionType {
   id: number;
   name: string;
 }
+
+export const initialRegionValue = {
+  id: 0,
+  name: '',
+};
